some further parallelization bugfixes
Former-commit-id: 51ab0de8bd8b0e1bf98c1fcf98cf442ae4442fa6 [formerly 8febc2ff107630fbac6c186781b77c83a6f0772e]
Former-commit-id: 8cae4f1d86f96916e998c5132285dcca3136a63e

+++ src/test/solver/SolverStats_test.cpp
@@ -41,10 +41,12 @@ BOOST_AUTO_TEST_CASE(SolverStats_ConstructionAndFunctionality_test) {
 	CFDSolver<2> solver(testConfiguration, testFlow);
 
 	// make solver stats object
-	if (boost::filesystem::is_directory(natriumTmpDir)) {
-		boost::filesystem::remove_all(natriumTmpDir);
+	if (is_MPI_rank_0()) {
+		if (boost::filesystem::is_directory(natriumTmpDir)) {
+			boost::filesystem::remove_all(natriumTmpDir);
+		}
+		boost::filesystem::create_directory(natriumTmpDir);
 	}
-	boost::filesystem::create_directory(natriumTmpDir);
 	SolverStats<2> stats(&solver, natriumTmpFile.c_str());
 	BOOST_CHECK(boost::filesystem::exists(natriumTmpFile));
